pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "norgesgruppen-ai-vision"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = [
    # Pinned to exact sandbox versions (Python 3.11, PyTorch 2.6.0+cu124)
    "ultralytics==8.1.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "onnxruntime==1.20.0",       # CPU locally; sandbox uses onnxruntime-gpu 1.20.0
    "opencv-python-headless==4.9.0.80",
    "albumentations==1.3.1",
    "Pillow==10.2.0",
    "numpy<2.0",                  # numpy 1.26.4 in sandbox
    "scipy==1.12.0",
    "scikit-learn==1.4.0",
    "pycocotools==2.0.7",
    "ensemble-boxes==1.0.9",
    "timm==0.9.12",
    "supervision==0.18.0",
    "safetensors==0.4.2",
    "google-cloud-storage==2.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "bandit",
    "pip-audit",
]

# ---------------------------------------------------------------------------
# Ruff (linting + formatting)
# ---------------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "F",    # Pyflakes: unused imports, undefined names
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "I",    # isort: import ordering
    "UP",   # pyupgrade: modernize syntax
    "B",    # flake8-bugbear: common bugs
    "SIM",  # flake8-simplify: simplifiable code
    "RUF",  # Ruff-specific rules
    "S",    # flake8-bandit: security checks
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # assert is expected in test files
"scripts/**" = ["UP017"]  # datetime.UTC requires Python 3.12+, container runs 3.11

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ---------------------------------------------------------------------------
# Mypy (type checking)
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "ultralytics.*"
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
addopts = "-q --tb=line"

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
show_missing = true
fail_under = 80
